Ban & massAction fix
When an admin bans a players on dashboard it kicks the player instead of banning him, if the player rejoin to the server it checks the player's id in db. iIf he/she has an active ban, the player'll be kicked.

massAction had typos and wasn't log the action to moderation sorry about that :( pls check again don't trust me :laughing: